Add functionality to merge windfarm and turbine locations (#5)
* Add windturbine to windfarm mapper

* Improve turbine to windfarm mapper

* Add rename rules and general use of io.readers

* Only write skipped turbines from austria in debug mode

* Add converter for turbines from sweden

* Update generation of euromap

* Fix typo

* Fix filter for swedish turbines

* Add support for processing TheWindPower.net data

* Add windfarms to osm_data_fetcher

* Add extended osm_data_fetcher

* Some more file handling

* Improve logging

* Refactor build_query function

* By default query for windturbine and windfarm data

* Restore optimization to compute lat/lon of ways

+++ json2tab/Turbine.py
@@ -28,7 +28,7 @@ class Turbine:
 
     height_offset: float = None
     wind_farm: str = None
-    n_turbines: float = None
+    n_turbines: int = None
     operator: str = None
 
     start_date: date = None
